// src/api/models/DsChatResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Mailing app
 * my api
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface DsChatResponse
 */
export interface DsChatResponse {
    /**
     * 
     * @type {number}
     * @memberof DsChatResponse
     */
    friends?: number;
    /**
     * 
     * @type {number}
     * @memberof DsChatResponse
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof DsChatResponse
     */
    img?: string;
    /**
     * 
     * @type {string}
     * @memberof DsChatResponse
     */
    info?: string;
    /**
     * 
     * @type {string}
     * @memberof DsChatResponse
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof DsChatResponse
     */
    nickname?: string;
    /**
     * 
     * @type {number}
     * @memberof DsChatResponse
     */
    subscribers?: number;
}

/**
 * Check if a given object implements the DsChatResponse interface.
 */
export function instanceOfDsChatResponse(value: object): value is DsChatResponse {
    return true;
}

export function DsChatResponseFromJSON(json: any): DsChatResponse {
    return DsChatResponseFromJSONTyped(json, false);
}

export function DsChatResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): DsChatResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'friends': json['friends'] == null ? undefined : json['friends'],
        'id': json['id'] == null ? undefined : json['id'],
        'img': json['img'] == null ? undefined : json['img'],
        'info': json['info'] == null ? undefined : json['info'],
        'name': json['name'] == null ? undefined : json['name'],
        'nickname': json['nickname'] == null ? undefined : json['nickname'],
        'subscribers': json['subscribers'] == null ? undefined : json['subscribers'],
    };
}

export function DsChatResponseToJSON(json: any): DsChatResponse {
    return DsChatResponseToJSONTyped(json, false);
}

export function DsChatResponseToJSONTyped(value?: DsChatResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'friends': value['friends'],
        'id': value['id'],
        'img': value['img'],
        'info': value['info'],
        'name': value['name'],
        'nickname': value['nickname'],
        'subscribers': value['subscribers'],
    };
}